backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
pdf2image==1.16.3
pytesseract==0.3.10
python-dotenv==1.0.0
pydantic==2.5.0
email-validator==2.1.0
aiofiles==23.2.0
PyMySQL==1.1.0
